refactor: move prepareOpenAPI functions to own directory
