+++ src/composables/useSidebar.ts
@@ -2,7 +2,7 @@ import type { OpenAPI, OpenAPIV3 } from '@scalar/openapi-types'
 import type { DefaultTheme } from 'vitepress'
 import { OpenApi } from '../lib/OpenApi'
 import { httpVerbs } from '../index'
-import { prepareOpenAPI } from '../lib/prepareOpenAPI'
+import { prepareOpenAPI } from '../lib/prepareOpenAPI/prepareOpenAPI'
 import { generateSidebarItemsByPaths } from '../lib/sidebar/generateSidebarItemsByPaths'
 import { cleanSidebarItems } from '../lib/sidebar/cleanSidebarItems'
 import { useOpenapi } from './useOpenapi'
