+++ src/lib/codeSamples/generateCodeSamplePhp.ts
@@ -7,7 +7,7 @@ export function generateCodeSamplePhp({ url, method, headers, body, query }: OAR
 
   phpCode += `$method = '${method}';\n`
 
-  if (Object.keys(headers).length > 0) {
+  if (headers && Object.keys(headers).length > 0) {
     phpCode += `$headers = [\n`
     for (const [key, value] of Object.entries(headers)) {
       phpCode += `    '${key}' => '${value}',\n`
@@ -15,7 +15,7 @@ export function generateCodeSamplePhp({ url, method, headers, body, query }: OAR
     phpCode += `];\n`
   }
 
-  if (Object.keys(query).length > 0) {
+  if (query && Object.keys(query).length > 0) {
     phpCode += `$query = http_build_query([\n`
     for (const [key, value] of Object.entries(query)) {
       phpCode += `    '${key}' => '${value}',\n`
@@ -28,11 +28,11 @@ export function generateCodeSamplePhp({ url, method, headers, body, query }: OAR
   }
 
   phpCode += `\n$ch = curl_init();\n`
-  phpCode += `curl_setopt($ch, CURLOPT_URL, $url${Object.keys(query).length ? ` . '?' . $query` : ''});\n`
+  phpCode += `curl_setopt($ch, CURLOPT_URL, $url${Object.keys(query ?? {}).length ? ` . '?' . $query` : ''});\n`
   phpCode += `curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n`
   phpCode += `curl_setopt($ch, CURLOPT_CUSTOMREQUEST, $method);\n`
 
-  if (Object.keys(headers).length > 0) {
+  if (headers && Object.keys(headers).length > 0) {
     phpCode += `curl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\n`
   }
